Update pipeline template model to support tags and version
- This breaks the old api in @register_pipeline
- WIP 'custom' pipeline registry model for emitting pipeline templates
