+++ pbsmrtpipe/__init__.py
@@ -1,4 +1,4 @@
-VERSION = (0, 19, 8)
+VERSION = (0, 19, 9)
 
 
 def get_version():
