+++ pbsmrtpipe/pb_pipelines/pb_pipelines_dev.py
@@ -8,7 +8,13 @@ from .pb_pipelines_sa3 import Constants
 log = logging.getLogger(__name__)
 
 
-@register_pipeline(to_pipeline_ns("dev_local"), "Dev Local Hello Pipeline")
+def dev_register(relative_id, display_name, tags=()):
+    pipeline_id = to_pipeline_ns(relative_id)
+    ptags = list(set(tags + ('dev', )))
+    return register_pipeline(pipeline_id, display_name, "0.1.0", tags=ptags)
+
+
+@dev_register("dev_local", "Dev Local Hello Pipeline", tags=("local", ))
 def get_dev_local_pipeline():
     """Simple example pipeline"""
     b1 = [('$entry:e_01', 'pbsmrtpipe.tasks.dev_hello_world:0')]
@@ -23,21 +29,21 @@ def get_dev_local_pipeline():
     return b1 + b2 + b3 + b4
 
 
-@register_pipeline("pbsmrtpipe.pipelines.dev_01", "Example Dev 01 pipeline")
+@dev_register("dev_01", "Example Dev 01 pipeline")
 def f():
     """Simplest possible pipeline, a single Task"""
     b = [("$entry:e_01", "pbsmrtpipe.tasks.dev_hello_world:0")]
     return b
 
 
-@register_pipeline("pbsmrtpipe.pipelines.dev_01_ds", "Example Dev 01 Subread DataSet pipeline")
+@dev_register("dev_01_ds", "Example Dev 01 Subread DataSet pipeline", tags=("report", ))
 def f():
     """Simplest possible pipeline, a single Task"""
     b = [("$entry:eid_subread", "pbsmrtpipe.tasks.dev_subread_report:0")]
     return b
 
 
-@register_pipeline("pbsmrtpipe.pipelines.dev_02", "Example Dev 02 pipeline")
+@dev_register("dev_02", "Example Dev 02 pipeline")
 def f():
     # Reuse existing pipeline and reference a specific task output
     # in the pipeline. The format is {pipeline_id}:{task_id}:{task_out_index}
@@ -51,7 +57,7 @@ def f():
     return b + b2
 
 
-@register_pipeline("pbsmrtpipe.pipelines.dev_03", "Example Dev 03 pipelines")
+@dev_register("dev_03", "Example Dev 03 pipelines", tags=('dev', ))
 def f():
     """Reuse of pipeline 02"""
     b = [("pbsmrtpipe.pipelines.dev_02:pbsmrtpipe.tasks.dev_hello_worlder:0", "pbsmrtpipe.tasks.dev_hello_garfield:0")]
@@ -61,7 +67,7 @@ def f():
     return b + b2
 
 
-@register_pipeline("pbsmrtpipe.pipelines.dev_04", "Example Dev 04")
+@dev_register("dev_04", "Example Dev 04")
 def f():
     b0 = [("pbsmrtpipe.pipelines.dev_03:pbsmrtpipe.tasks.dev_hello_garfield:0", "pbsmrtpipe.tasks.dev_txt_to_fofn:0")]
     b1 = [("pbsmrtpipe.pipelines.dev_03:pbsmrtpipe.tasks.dev_hello_garfield:0", "pbsmrtpipe.tasks.dev_txt_to_fasta:0")]
@@ -74,7 +80,7 @@ def f():
     return b0 + b1 + b3
 
 
-@register_pipeline("pbsmrtpipe.pipelines.dev_04_w_static_task", "Pipeline that leverages a python static tasks")
+@dev_register("dev_04_w_static_task", "Pipeline that leverages a python static tasks")
 def f():
 
     # pbsmrtpipe.tasks.dev_static_txt_task needs to be ported
@@ -92,7 +98,7 @@ def f():
     return b1 + b2 + b3
 
 
-@register_pipeline(to_pipeline_ns("dev_local_chunk"), "Dev Local Hello Chunkable Pipeline")
+@dev_register("dev_local_chunk", "Dev Local Hello Chunkable Pipeline", tags=("local", "chunk"))
 def get_dev_local_chunk():
     """Simple example pipeline"""
     b1 = [("$entry:e_01", "pbsmrtpipe.tasks.dev_txt_to_fofn:0")]
@@ -110,7 +116,7 @@ def get_dev_local_chunk():
     return b1 + b3 + b4 + b5
 
 
-@register_pipeline(to_pipeline_ns("dev_local_fasta_chunk"), "Dev local Task for Chunking pipelines")
+@dev_register("dev_local_fasta_chunk", "Dev local Task for Chunking pipelines", tags=('chunking',))
 def get_dev_local_pipeline_chunk():
 
     b1 = [("pbsmrtpipe.pipelines.dev_03:pbsmrtpipe.tasks.dev_hello_garfield:0", "pbsmrtpipe.tasks.dev_txt_to_fasta:0")]
@@ -122,7 +128,7 @@ def get_dev_local_pipeline_chunk():
     return b1 + b2 + b3
 
 
-@register_pipeline(to_pipeline_ns("dev_dist"), "Dev Hello Distributed Workflow Pipeline")
+@dev_register("dev_dist", "Dev Hello Distributed Workflow Pipeline", tags=('cluster', ))
 def get_dist_dev_pipeline():
     """Simple distributed example pipeline"""
     bs = get_dev_local_pipeline()
@@ -133,7 +139,7 @@ def get_dist_dev_pipeline():
     return bs + b2
 
 
-@register_pipeline(to_pipeline_ns("dev_05"), "Reference Set Report")
+@dev_register("dev_05", "Reference Set Report", tags=('reports', "reference" ))
 def get_reference_ds_report():
     """Generate a simple report and plot from Reference DataSet"""
